CS335 object models The basketball goal had to be broken into parts since the model was too large for the obj loader method.
The texture coordinates did not be applied for some reason to the obj exports, but I have included the textures that are associated with the given objects.

There was no way to reduce a ball object to a small enough vertices count to export, so I recommend just using the sphere primitive already provided to us by Dr. Yang.  I did include the ball objects texture though.  It's extremely simple.